Updated mode to conform better to the Munix SpaceWarp template. Added capability to Add Node if there isn't one.
