+++ Debug/BepInEx/plugins/maneuver_node_controller/LICENSE
@@ -1,0 +1,1 @@
+INSERT YOUR LICENSE HERE
