+++ Debug/BepInEx/plugins/maneuver_node_controller/README.txt
@@ -1,0 +1,2 @@
+# Maneuver Node Controller
+Provides an interface to finely tune maneuver nodes
