+++ Debug/BepInEx/plugins/maneuver_node_controller/swinfo.json
@@ -1,0 +1,22 @@
+{
+  "mod_id": "maneuver_node_controller",
+  "author": "schlosrat",
+  "name": "Maneuver Node Controller",
+  "description": "Provides an interface to finely tune maneuver nodes",
+  "source": "https://github.com/schlosrat/ManeuverNodeController",
+  "version": "0.6.0",
+  "version_check": "",
+  "dependencies": [
+    {
+      "id":  "SpaceWarp",
+      "version": {
+        "min": "1.0.1",
+        "max": "*"
+      }
+    }
+  ],
+  "ksp2_version": {
+    "min": "0.1.0",
+    "max": "*"
+  }
+}
