+++ ManeuverNodeController.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.2.32630.192
+VisualStudioVersion = 17.5.33502.453
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ManeuverNodeController", "ManeuverNodeController.csproj", "{D9219E58-4E3C-4297-8C62-8E55EEDF1C84}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ManeuverNodeController", "ManeuverNodeControllerProject\ManeuverNodeController.csproj", "{1FABA8FB-7AA7-453E-A6FD-BE014A4C6833}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,12 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D9219E58-4E3C-4297-8C62-8E55EEDF1C84}.Debug|Any CPU.ActiveCfg = Release|Any CPU
-		{D9219E58-4E3C-4297-8C62-8E55EEDF1C84}.Debug|Any CPU.Build.0 = Release|Any CPU
-		{D9219E58-4E3C-4297-8C62-8E55EEDF1C84}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D9219E58-4E3C-4297-8C62-8E55EEDF1C84}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1FABA8FB-7AA7-453E-A6FD-BE014A4C6833}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1FABA8FB-7AA7-453E-A6FD-BE014A4C6833}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1FABA8FB-7AA7-453E-A6FD-BE014A4C6833}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1FABA8FB-7AA7-453E-A6FD-BE014A4C6833}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {1256F93C-8A09-4A56-9C77-ED68416842F1}
 	EndGlobalSection
 EndGlobal
